+++ src/parse-text.test.ts
@@ -6,15 +6,15 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 import * as fs from 'fs';
-import { parseCptTextInternal } from './parse-text';
+import { parsePaletteTextInternal } from './parse-text';
 
 const PATH = __dirname + '/../fixtures';
 const FILENAMES = fs.readdirSync(PATH).filter(x => x.endsWith('.cpt'));
 
-describe('parseCptTextInternal', () => {
+describe('parsePaletteTextInternal', () => {
   it.each(FILENAMES)('returns a color palette', (filename) => {
     const text = fs.readFileSync(PATH + '/' + filename).toString();
-    const actual = parseCptTextInternal(text);
+    const actual = parsePaletteTextInternal(text);
     // fs.writeFileSync(PATH + '/' + filename.replace('.cpt', '.json'), JSON.stringify(actual));
 
     const expected = JSON.parse(fs.readFileSync(PATH + '/' + filename.replace('.cpt', '.json')).toString());